Refactor Folder Structure
All component files in same folder
