+++ client/app/components/App/app.component.spec.ts
@@ -1,5 +1,5 @@
 import { TestBed } from '@angular/core/testing';
-import { AppComponent } from '../components/app.component';
+import { AppComponent } from './app.component';
 describe('App', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({ declarations: [AppComponent]});
