+++ client/app/components/App/app.component.ts
@@ -2,8 +2,8 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'my-app',
-  templateUrl: '../templates/app.component.html',
-  styleUrls: ['../styles/app.component.scss'],
+  templateUrl: './app.component.html',
+  styleUrls: ['./app.component.scss'],
 })
 
 export class AppComponent {
