+++ client/app/components/DayCard/dayCard.component.scss
