+++ client/app/modules/app.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { AppComponent } from '../components/app.component';
-import { DayCardComponent } from '../components/dayCard.component';
+import { AppComponent } from '../components/App/app.component';
+import { DayCardComponent } from '../components/DayCard/dayCard.component';
 
 @NgModule({
   imports: [
